VAE Model sort of finished.
If trained on one label, it converges to near 0 regardless of the number of layer used. However, when two labels are used, the training does not reach zero even for a bit. Additionally, for two labels, only 3 layers can be used (input, latent, output). If more hidden layers are introduced, it will only reconstruct one of the label, not both